#8308 Moved paramActions variable into utils
